Fix an IntLongMath issue
In common module, GridUtils.getWaveLengthIndexFromSpectrumAssigment()
method does not need to return a long.
Keeping an int is sufficient and thus avoids the IntLongMath issue
reported by ErrorProne.

+++ common/src/main/java/org/opendaylight/transportpce/common/fixedflex/GridUtils.java
@@ -62,7 +62,7 @@ public final class GridUtils {
      * @param index int
      * @return the wavelength number.
      */
-    public static long getWaveLengthIndexFromSpectrumAssigment(int index) {
+    public static int getWaveLengthIndexFromSpectrumAssigment(int index) {
         return (GridConstant.EFFECTIVE_BITS - index) / GridConstant.NB_SLOTS_100G;
     }
